+++ GUI/insert.py
@@ -2,12 +2,13 @@ from tkinter import *
 import os
 import sys
 from PIL import ImageTk, Image
-import threading
 import subprocess
+import time
+
+
 
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-from filehandling import initial_card_check
-
+from filehandling import CARD_PATH, KEY, read_card
 #Windows specifications
 window = Tk()
 
@@ -36,36 +37,100 @@ photo_label = Label(window, bg="#E7E6DD")
 photo = ImageTk.PhotoImage(image)
 photo_label.config(image=photo)
 photo_label.place(x=350, y=20)
+does_exist = False
 
+def check_removable_drive():
+    # Run system command to get drive information
+    cmd = "wmic logicaldisk get caption, drivetype"
+    output = subprocess.check_output(cmd, shell=True).decode("utf-8")
+
+    # Parse the output to extract removable drives
+    drives = []
+    lines = output.strip().split("\n")[1:]
+    for line in lines:
+        drive, drivetype = line.split()
+        if int(drivetype) == 2:
+            drives.append(drive)
+
+    return drives
+
+def get_removable_drive_path(drive_letter):
+    cmd = "wmic logicaldisk where caption='{}' get name".format(drive_letter)
+    output = subprocess.check_output(cmd, shell=True).decode("utf-8")
+    lines = output.strip().split("\n")
+    if len(lines) >= 2:
+        drive_path = lines[1].strip()
+        return drive_path
+    return None
+
+
+def initial_card_check():
+    # Initial check for removable drives
+    global CARD_PATH
+
+    previous_drives = check_removable_drive()
+    print("Please insert Flash Drive...")
+
+    # Wait for a short interval before checking again
+    interval = 0.5  # in seconds
+    time.sleep(interval)
+
+    # Check for removable drives
+    current_drives = check_removable_drive()
+
+    # Compare current and previous drives to detect changes
+    added_drives = [drive for drive in current_drives if drive not in previous_drives]
+    removed_drives = [drive for drive in previous_drives if drive not in current_drives]
+
+    # Handle added and removed drives
+    if added_drives:
+        print("Card Inserted")
+        print("Validating...")
+        drive_path = get_removable_drive_path(added_drives[0])
+        file_path = f"{drive_path}\\record.txt"
+      
+        CARD_PATH = file_path 
+        if not os.path.isfile(file_path):
+            print("file not found")
+            return 2
+        else:   
+            return 1
+
+    if removed_drives:
+        print("Removed Card")
+        return 3
+        
+
+    # Update the previous drives list
+    previous_drives = current_drives
 
 def after_init():
-    # Actions to perform after the window is initialized
-    print("After init")
+    global does_exist
     # Call the initial_card_check function
-    does_exist = False
     # Process the result
-    while True:
-        result = initial_card_check()
-        if result:
-            print("test")
-            does_exist = True
-            break
+    result = initial_card_check()
+    if result == 1:
+        does_exist = True
     
     if does_exist:
-        window.quit()
+        acc = read_card(CARD_PATH, KEY)
+        print(KEY)
+        window.destroy()
         current_directory = os.path.dirname(os.path.abspath(__file__))
         script_path = os.path.join(current_directory, "log_in.py")
-        subprocess.run(["python", script_path]) 
+        subprocess.run(["python", script_path])
+
+    if result == 2:
+        window.destroy()
+        current_directory = os.path.dirname(os.path.abspath(__file__))
+        script_path = os.path.join(current_directory, "register.py")
+        subprocess.run(["python", script_path])
+    
+    window.after(100, after_init)
 
        
-
-       
-
-
-thread = threading.Thread(target=after_init)
-thread.start()
-
-
+    
+after_init()
 
 window.mainloop()
