+++ README.md
@@ -1,10 +1,13 @@
 # Bank-Python
 Project for ProgLang
 
-Prerequisites
-In order to run our application you need to run these following commands or install packages like pillow and psutil
+## Prerequisites
+In order to run our application you need to run these following commands or install packages like pillow and psutil also tkcalendar
 ```python
     pip install pillow
+```
+```python
+    pip install psutil
 ```
 ```python
     pip install psutil
