+++ test/tempCodeRunnerFile.py
@@ -1,0 +1,1 @@
+path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
